+++ app/components/bulma/notification_component.rb
@@ -2,13 +2,18 @@
 
 module Bulma
   class NotificationComponent < Component
+    renders_one :delete, DeleteComponent
+
     def initialize(color: nil, light: false)
       @color = color
       @light = light
     end
 
     def call
-      content_tag(:div, content, class: classes)
+      content_tag(:div, class: classes) do
+        concat delete if delete?
+        concat content
+      end
     end
 
     def classes